+++ SDK/multiplayer/src/PlatformUtils.win.cpp
@@ -11,7 +11,7 @@
  * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Software.
  *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR 
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR f
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, 
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE 
  * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER 
@@ -215,7 +215,7 @@ void	ShowProgressMessage(const char * inMsg, float * inProgress)
 		ShowWindow(prog,SW_SHOWNORMAL);
 		if (v >= 0.0)
 		{
-			int n = v * 1000.0;
+			int n = (int) v * 1000.0;
 //			SendMessage(prog, PBM_SETMARQUEE, (WPARAM) 0, 0); 		
 			SendMessage(prog, PBM_SETPOS, (WPARAM) n, 0); 		
 		} else {
